# Check View sizes in reallocWithoutInitializing