In https://github.com/arborx/ArborX/pull/538, it turned out to be useful to avoid allocating memory in `reallocWithoutInitializing` when the sizes already fit allowing to resize `Kokkos::Views` used in `query` beforehand and thus avoiding some synchronizations.