@aprokop 2021-07-27
> In #538, it turned out to be useful to avoid allocating memory in `reallocWithoutInitializing` when the sizes already fit allowing to resize `Kokkos::Views` used in `query` beforehand and thus avoiding some synchronizations.

Could you please provide more details? From the patch I see you do no-op, if one tries to realloc to the current size (I think this is what you mean by "fit"). It seems like a reasonable shortcut. But I have a hard time figuring out when this happens, and how common that is.