@masterleinad 2021-07-27
> Could you please provide more details? From the patch I see you do no-op, if one tries to realloc to the current size (I think this is what you mean by "fit"). It seems like a reasonable shortcut. But I have a hard time figuring out when this happens, and how common that is.

Yes, exactly, this pull request makes reallocating with the current size a no-op. It's clearly not a common case but if the user knows the correct size for `offsets` and `values` already, it avoids reallocations (and thus more synchronizations). This was part of me figuring out how to run multiple streams concurrently using nearest searches (before switching to callbacks and intersection searches in that pull request). It's a small optimization.
BTW, `Kokkos::realloc` is supposed to get a `WithoutInitialzing` option that we should then use instead after the next release (and I consider suggesting the same patch there, too).