@masterleinad 2021-07-29
> It seems like there's a performance regression in Serial construction ([log](https://github.com/masterleinad/ArborX/runs/3175782026)) for a small problem.

This is weird. Running
```
#include <details/ArborX_DetailsUtils.hpp>

#include <benchmark/benchmark.h>

template <class MemorySpace>
void BM_generic(benchmark::State &state) {
  int n          = state.range(0);
  Kokkos::View<double*, MemorySpace> view("dummy_view", n);
  for (auto _ : state) {
    ArborX::reallocWithoutInitializing(view, n);
  }
}

#define REGISTER_BENCHMARK(SPACE)                     \
  BENCHMARK_TEMPLATE(BM_generic, SPACE)               \
      ->Ranges({{1, 8 << 16}}) \
      ->UseRealTime();

#ifdef KOKKOS_ENABLE_SERIAL
REGISTER_BENCHMARK(Kokkos::HostSpace);
#endif
#ifdef KOKKOS_ENABLE_CUDA
REGISTER_BENCHMARK(Kokkos::CudaSpace);
#endif
#ifdef KOKKOS_ENABLE_HIP
REGISTER_BENCHMARK(Kokkos::Experimental::HIPSpace);
#endif
#ifdef KOKKOS_ENABLE_SYCL
REGISTER_BENCHMARK(Kokkos::Experimental::SYCLDeviceUSMSpace);
REGISTER_BENCHMARK(Kokkos::Experimental::SYCLSharedUSMSpace);
#endif

int main(int argc, char **argv) {
  Kokkos::initialize(argc, argv);
  benchmark::Initialize(&argc, argv);
  benchmark::RunSpecifiedBenchmarks();
  Kokkos::finalize();
}
```
I see
```
BM_generic<Kokkos::HostSpace>/1/real_time            1397 ns         1396 ns       458790
BM_generic<Kokkos::HostSpace>/8/real_time            1761 ns         1761 ns       481464
BM_generic<Kokkos::HostSpace>/64/real_time           1969 ns         1969 ns       295956
BM_generic<Kokkos::HostSpace>/512/real_time          2024 ns         2024 ns       348890
BM_generic<Kokkos::HostSpace>/4096/real_time         2845 ns         2844 ns       260350
BM_generic<Kokkos::HostSpace>/32768/real_time       79821 ns        79818 ns        11276
BM_generic<Kokkos::HostSpace>/262144/real_time     270133 ns       270103 ns         2510
BM_generic<Kokkos::HostSpace>/524288/real_time     508516 ns       508489 ns         1305
```
before and 
```
-----------------------------------------------------------------------------------------
Benchmark                                               Time             CPU   Iterations
-----------------------------------------------------------------------------------------
BM_generic<Kokkos::HostSpace>/1/real_time            18.0 ns         17.9 ns     44694859
BM_generic<Kokkos::HostSpace>/8/real_time            17.7 ns         17.7 ns     38384352
BM_generic<Kokkos::HostSpace>/64/real_time           16.1 ns         16.1 ns     42493309
BM_generic<Kokkos::HostSpace>/512/real_time          16.3 ns         16.3 ns     42682222
BM_generic<Kokkos::HostSpace>/4096/real_time         16.4 ns         16.4 ns     41885972
BM_generic<Kokkos::HostSpace>/32768/real_time        16.3 ns         16.3 ns     41934500
BM_generic<Kokkos::HostSpace>/262144/real_time       19.9 ns         19.9 ns     42274479
BM_generic<Kokkos::HostSpace>/524288/real_time       16.3 ns         16.3 ns     43367300
```
after this patch on my laptop.