@masterleinad 2021-08-02
Running the `bvh_driver` manually on Summit, I could also not notice any significant change in performance.

Retest this please.