@masterleinad 2021-08-03
The corresponding `Kokkos` pull request is https://github.com/kokkos/kokkos/pull/4207.