@masterleinad 2021-08-03
The new results look more reasonable: https://code.ornl.gov/ecpcitest/alexa/-/jobs/1001696.