@aprokop 2021-08-03
Merged with only SYCL tests failing due to certificate fetching issue.